entrypoints 整理
